+++ frontend/src/App.vue
@@ -232,7 +232,7 @@ onMounted(async () => {
     </div>
     <header class="topbar">
       <div class="tb-left">
-        <img class="tb-logo" :src="logoUrl" alt="甲骨易" width="28" height="28" />
+        <img class="tb-logo" :src="logoUrl" alt="甲骨易" width="42" height="42" />
         <div class="tb-title"><b>甲骨易</b> 经营看板</div>
         <PeriodPicker />
       </div>